+++ src/service.js
@@ -11,14 +11,14 @@ function createDragHandler({ctx, name, drake}) {
 }
 
 export class DragulaService {
-  constructor ({name, eventBus, bags, drake, options}) {
+  constructor ({name, eventBus, drakes, options}) {
     this.options = options || {}
     this.logging = options.logging
     this.name = name
-    this.bags = bags = {} // bag store
+    this.drakes = drakes = {} // drake store
     this.eventBus = eventBus
-    this.drake = drake
     this.createDragHandler = options.createDragHandler || createDragHandler
+
     this.events = [
       'cancel',
       'cloned',
@@ -44,40 +44,46 @@ export class DragulaService {
     throw new Error(msg)
   }
 
-  get bagNames() {
-    return Object.keys(this.bags)
+  _validate(method, name) {
+    if (!name) {
+      this.error(`${method} must take a drake name as the first argument`)
+    }
+  }
+
+  get drakeNames() {
+    return Object.keys(this.drakes)
   }
 
   add (name, drake) {
-    drake = drake || this.drake
-    this.log('add (bag)', name, drake)
-    let bag = this.find(name)
-    if (bag) {
-      this.log('existing bags', this.bagNames)
-      let errMsg = `Bag named: "${name}" already exists for this service [${this.name}]. 
+    this.log('add (drake)', name, drake)
+    this._validate('add', name)
+    if (this.find(name)) {
+      this.log('existing drakes', this.drakeNames)
+      let errMsg = `Drake named: "${name}" already exists for this service [${this.name}]. 
       Most likely this error in cause by a race condition evaluating multiple template elements with 
-      the v-dragula directive having the same bag name. Please initialise the bag in the created() life cycle hook of the VM to fix this problem.`
+      the v-dragula directive having the same drake name. Please initialise the drake in the created() life cycle hook of the VM to fix this problem.`
       this.error(msg)
     }
-    this.bags[name] = drake
+
+    this.drakes[name] = drake
     if (drake.models) {
       this.handleModels(name, drake)
     }
-    if (!bag.initEvents) {
-      this.setupEvents(name, bag)
+    if (!drake.initEvents) {
+      this.setupEvents(name, drake)
     }
-    return bag
+    return drake
   }
 
   find (name) {
-    this.log('find (bag) by name', name)
-    return this.bags[name]
+    this.log('find (drake) by name', name)
+    this._validate('find', name)
+    return this.drakes[name]
   }
 
   handleModels (name, drake) {
-    drake = drake || this.drake
     this.log('handleModels', name, drake)
-
+    this._validate('handleModels', name)
     if (drake.registered) { // do not register events twice
       return
     }
@@ -103,37 +109,36 @@ export class DragulaService {
 
   destroy (name) {
     this.log('destroy', name)
-    let bag = this.find(name)
-    if (!bag) { return }
-    bag.destroy()
-    this.delete(name)
+    this._validate('destroy', name)
+    let drake = this.find(name)
+    if (!drake) { return }
+    drake.destroy()
+    this._delete(name)
   }
 
-  delete(name) {
-    delete this.bags[name]
+  _delete(name) {
+    delete this.drakes[name]
   }
 
   setOptions (name, options) {
     this.log('setOptions', name, options)
-    if (!name) {
-      console.error('setOptions must take the name of the bag to set options for')
-      return this
-    }
-    let bag = this.add(name, dragula(options))
-    this.handleModels(name, bag)
+    this._validate('setOptions', name)
+    let drake = this.add(name, dragula(options))
+    this.handleModels(name, drake)
     return this
   }
 
-  setupEvents (name, bag) {
-    this.log('setupEvents', name, bag)
-    bag.initEvents = true
+  setupEvents (name, drake) {
+    this.log('setupEvents', name, drake)
+    this._validate('setupEvents', name)
+    drake.initEvents = true
     let _this = this
     let emitter = type => {
       function replicate () {
         let args = Array.prototype.slice.call(arguments)
         _this.eventBus.$emit(type, [name].concat(args))
       }
-      bag.on(type, replicate)
+      drake.on(type, replicate)
     }
     this.events.forEach(emitter)
   }
@@ -146,13 +151,11 @@ export class DragulaService {
   }
 
   findModelForContainer (container, drake) {
-    drake = drake || this.drake
     this.log('findModelForContainer', container, drake)
     return (this.findModelContainerByContainer(container, drake) || {}).model
   }
 
   findModelContainerByContainer (container, drake) {
-    drake = drake || this.drake
     if (!drake.models) {
       return
     }